. 修正 cmp，setkey 函数。避免动画背景残留。改进分区探测。

+++ stage2/shared.h
@@ -1704,8 +1704,8 @@ void grub_halt (void); //__attribute__ ((noreturn));
 /* The key map.  */
 struct key_map
 {
-	unsigned short from_code;
-	unsigned short to_code;
+	unsigned int from_code;
+	unsigned int to_code;
 };
 extern struct key_map ascii_key_map[];
 extern int remap_ascii_char (int key);
